+++ warp/main.go
@@ -5,23 +5,20 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/cloudogu/k8s-registry-lib/dogu"
-	"github.com/cloudogu/k8s-registry-lib/repository"
+	componentv1 "github.com/cloudogu/k8s-component-lib/api/v1"
+	warpmenu "github.com/cloudogu/k8s-warp-menu-entry-lib/api/v1"
 	"github.com/cloudogu/warp-assets/config"
 	warpCtrl "github.com/cloudogu/warp-assets/controller"
 	"github.com/cloudogu/warp-assets/logging"
 	"k8s.io/apimachinery/pkg/runtime"
+	k8sTypes "k8s.io/apimachinery/pkg/types"
 	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
-	"k8s.io/client-go/kubernetes"
 	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
-	"k8s.io/client-go/rest"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/cache"
 	"sigs.k8s.io/controller-runtime/pkg/manager"
 	"sigs.k8s.io/controller-runtime/pkg/metrics/server"
 	"sigs.k8s.io/controller-runtime/pkg/webhook"
-
-	doguv2 "github.com/cloudogu/k8s-dogu-lib/v2/api/v2"
 )
 
 var (
@@ -38,7 +35,8 @@ type k8sManager interface {
 
 func init() {
 	utilruntime.Must(clientgoscheme.AddToScheme(scheme))
-	utilruntime.Must(doguv2.AddToScheme(scheme))
+	utilruntime.Must(warpmenu.AddToScheme(scheme))
+	utilruntime.Must(componentv1.AddToScheme(scheme))
 	// +kubebuilder:scaffold:scheme
 
 	if err := logging.ConfigureLogger(); err != nil {
@@ -80,28 +78,21 @@ func start() error {
 }
 
 func setupWarpMenuReconciler(warpMenuManager k8sManager, watchNamespace string) error {
-	clientset, err := getK8sClientSet(warpMenuManager.GetConfig())
-	if err != nil {
-		return fmt.Errorf("create k8s client set: %w", err)
-	}
+	warpMenuClient := warpMenuManager.GetClient()
 
-	client := warpMenuManager.GetClient()
-	configMapInterface := clientset.CoreV1().ConfigMaps(watchNamespace)
-	globalConfigRepo := repository.NewGlobalConfigRepository(configMapInterface)
-	doguVersionRegistry := dogu.NewDoguVersionRegistry(configMapInterface)
-	localDoguRepo := dogu.NewLocalDoguDescriptorRepository(configMapInterface)
-
-	deploymentName, err := config.ReadDeploymentName()
+	componentName, err := config.ReadComponentName()
 	if err != nil {
 		return fmt.Errorf("read config value 'deployment name': %w", err)
 	}
-	eventRecorder := warpMenuManager.GetEventRecorderFor(deploymentName)
+	eventRecorder := warpMenuManager.GetEventRecorder(componentName)
 
 	warpMenuPath, err := config.ReadWarpPath()
 	if err != nil {
 		return fmt.Errorf("read config value 'warp path': %w", err)
 	}
-	reconciler := warpCtrl.NewWarpMenuReconciler(client, globalConfigRepo, doguVersionRegistry, localDoguRepo, eventRecorder, warpMenuPath, deploymentName)
+
+	componentCRKey := k8sTypes.NamespacedName{Name: componentName, Namespace: watchNamespace}
+	reconciler := warpCtrl.NewWarpMenuReconciler(warpMenuClient, eventRecorder, warpMenuPath, componentCRKey)
 	err = reconciler.SetupWithManager(warpMenuManager)
 	if err != nil {
 		return fmt.Errorf("setup reconciler with manager: %w", err)
@@ -119,15 +110,6 @@ func startManager(k8sManager k8sManager) error {
 	}
 
 	return nil
-}
-
-func getK8sClientSet(config *rest.Config) (*kubernetes.Clientset, error) {
-	k8sClientSet, err := kubernetes.NewForConfig(config)
-	if err != nil {
-		return nil, fmt.Errorf("create k8s client set: %w", err)
-	}
-
-	return k8sClientSet, nil
 }
 
 func getK8sManagerOptions(watchNamespace string) manager.Options {