+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [#21] Build WarpMenu based on WarpMenuEntry CR
+- [#21] Define default categories and entries for WarpMenu in values.yaml
+
+### Changed
+- [#24] Update Warp-Manu to v2.1.0 with support for localization of warp-menu entries
+- [#21] Write events for warp menu to the Component CR
 
 ## [v2.0.2] - 2026-03-12
 ### Changed